+++ src/actionCreator.spec.ts
@@ -92,52 +92,3 @@ test('isActionOfType', () => {
     isActionOfType(actionCreatorWithoutPayload, actionCreatorWithoutPayload())
   ).toBe(true);
 });
-
-// Type assertions
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-// const typeChecks = [
-//   // accepts results from action creators
-//   isValidRxBeachAction(actionCreatorWithoutPayload()),
-//   isValidRxBeachAction(actionCreatorWithPayload('test')),
-
-//   // should accept any value as argument
-//   isValidRxBeachAction(undefined),
-//   isValidRxBeachAction(null),
-//   isValidRxBeachAction({}),
-
-//   // @ts-expect-error: first argument of isActionOfType should be an action creator
-//   isActionOfType({}, actionCreatorWithoutPayload()),
-//   // @ts-expect-error: first argument of isActionOfType should be an action creator
-//   isActionOfType(1, actionCreatorWithoutPayload()),
-//   // @ts-expect-error: first argument of isActionOfType should be an action creator
-//   isActionOfType(null, actionCreatorWithoutPayload()),
-//   // @ts-expect-error: first argument of isActionOfType should be an action creator
-//   isActionOfType(undefined, actionCreatorWithoutPayload()),
-
-//   // second argument of isActionOfType accepts anything
-//   isActionOfType(actionCreatorWithoutPayload, actionCreatorWithoutPayload()),
-//   isActionOfType(actionCreatorWithoutPayload, actionCreatorWithPayload('test')),
-//   isActionOfType(actionCreatorWithPayload, actionCreatorWithoutPayload()),
-//   isActionOfType(actionCreatorWithPayload, actionCreatorWithPayload('test')),
-//   isActionOfType(actionCreatorWithPayload, undefined),
-//   isActionOfType(actionCreatorWithPayload, null),
-//   isActionOfType(actionCreatorWithPayload, {}),
-// ];
-
-const actionPairs = [
-  [actionCreatorWithPayload, actionCreatorWithPayload('asd')] as const,
-  [actionCreatorWithoutPayload, actionCreatorWithoutPayload()] as const,
-];
-
-for (const [creatorFn, anAction] of actionPairs) {
-  // @ts-expect-error Don't know how to fix this case...
-  isActionOfType(creatorFn, anAction);
-}
-
-const action1 = actionCreatorWithoutPayload();
-if (isActionOfType(actionCreatorWithoutPayload, action1)) {
-  // @ts-expect-error Assert that ActionWithoutPayload does not have a payload
-  // eslint-disable-next-line no-unused-expressions
-  action1.payload;
-}
