+++ src/internal/index.ts
@@ -1,11 +1,11 @@
 export {
+  RoutineFunc,
   VoidPayload,
   UnknownAction,
   ActionCreatorCommon,
   UnknownActionCreator,
   UnknownActionCreatorWithPayload,
 } from './types';
-export { RoutineFunc } from './routineFunc';
 export { coldMergeOperators } from '../operators/mergeOperators';
 export { defaultErrorSubject } from './defaultErrorSubject';
 export { rethrowErrorGlobally } from './rethrowErrorGlobally';
